MTA-725 created plan controller and unit tests
